*refactoring of ConvertModNode.
Move ConvertModNode from OnnxRuntime.h into ConvertModNode.cpp.
